Moved from chrome.processes API to chrome.tabs
The chrome.processes API is currently dev-only and is therefore not
preactically useable.

Using the chrome.tabs API requires more code and CPU-cycles at runtime.
It further gives less control/insight in where/what has caused a
problem. It does allow for almost equal fucntionality, so for now it is
considered good enough.